requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
asttokens==2.4.1
decorator==5.1.1
Django==4.2.14
exceptiongroup==1.2.2
executing==2.0.1
graphene==3.3
graphene-django==3.2.2
graphql-core==3.2.3
graphql-relay==3.2.0
ipython==8.18.1
jedi==0.19.1
kafka-python==2.0.2
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
promise==2.3
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
text-unidecode==1.3
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13

# Coding Standards
black==23.3.0
isort==5.10.1
flake8==6.0.0
autoflake==2.2.0
